refactor: extract SnapshotTransfer as an interface
